requirements.txt
fastapi==0.65.1
uvicorn==0.16.0
pytest==6.2.5
requests==2.27.1
pydantic==1.9.0
python-dotenv==0.19.2
tensorflow==2.7.0
boto3==1.20.27
sentencepiece==0.1.96
numpy==1.21.5
pandas==1.3.5
scipy==1.7.3